+++ app/src/main/java/com/gemini/energy/service/device/plugload/HotFoodCabinet.kt
@@ -30,9 +30,9 @@ class HotFoodCabinet(private val computable: Computable<*>, utilityRateGas: Util
     /**
      * Usage Hours
      * */
-    private var peakHours = 0
-    private var partPeakHours = 0
-    private var offPeakHours = 0
+    private var peakHours = 0.0
+    private var partPeakHours = 0.0
+    private var offPeakHours = 0.0
     private var usageHours: UsageSimple? = null
 
     private var cabinetVolume = 0.0
@@ -41,9 +41,9 @@ class HotFoodCabinet(private val computable: Computable<*>, utilityRateGas: Util
 
     override fun setup() {
 
-        peakHours = featureData["Peak Hours"]!! as Int
-        partPeakHours = featureData["Part Peak Hours"]!! as Int
-        offPeakHours = featureData["Off Peak Hours"]!! as Int
+        peakHours = featureData["Peak Hours"]!! as Double
+        partPeakHours = featureData["Part Peak Hours"]!! as Double
+        offPeakHours = featureData["Off Peak Hours"]!! as Double
         usageHours = UsageSimple(peakHours, partPeakHours, offPeakHours)
 
         cabinetVolume = featureData["Cabinet Volume"]!! as Double
