+++ app/src/main/java/com/gemini/energy/service/device/plugload/PreRinseSpray.kt
@@ -30,9 +30,9 @@ class PreRinseSpray(private val computable: Computable<*>, utilityRateGas: Utili
     /**
      * Usage Hours
      * */
-    private var peakHours = 0
-    private var partPeakHours = 0
-    private var offPeakHours = 0
+    private var peakHours = 0.0
+    private var partPeakHours = 0.0
+    private var offPeakHours = 0.0
     private var usageHours: UsageSimple? = null
 
     private var flowRate = 0.0
@@ -43,9 +43,9 @@ class PreRinseSpray(private val computable: Computable<*>, utilityRateGas: Utili
 
     override fun setup() {
 
-        peakHours = featureData["Peak Hours"]!! as Int
-        partPeakHours = featureData["Part Peak Hours"]!! as Int
-        offPeakHours = featureData["Off Peak Hours"]!! as Int
+        peakHours = featureData["Peak Hours"]!! as Double
+        partPeakHours = featureData["Part Peak Hours"]!! as Double
+        offPeakHours = featureData["Off Peak Hours"]!! as Double
         usageHours = UsageSimple(peakHours, partPeakHours, offPeakHours)
 
         flowRate = featureData["Flow Rate"]!! as Double
